support for indicating snapshot type (automated) and retaining individual snapshots indefinitely regardles of retention policy

+++ ebs-snapshot-manager.py
@@ -28,13 +28,27 @@ def lambda_handler(event, context):
 
 
     delete_on = datetime.date.today().strftime('%Y-%m-%d')
+        # limit snapshots to process to ones marked for deletion on this day
+        # AND limit snapshots to process to ones that are automated only
+        # FUTURE: AND exclude automated snapshots marked for permanent retention
+            #         { 'Name': 'tag:KeepForever', 'Values': ['Yes'] },
     filters = [
-        {'Name': 'tag-key', 'Values': ['DeleteOn']},
-        {'Name': 'tag-value', 'Values': [delete_on]},
+        { 'Name': 'tag:DeleteOn', 'Values': [delete_on] },
+        { 'Name': 'tag:Type', 'Values': ['Automated'] },
     ]
     snapshot_response = ec.describe_snapshots(OwnerIds=account_ids, Filters=filters)
 
-
     for snap in snapshot_response['Snapshots']:
-        print "Deleting snapshot %s" % snap['SnapshotId']
-        ec.delete_snapshot(SnapshotId=snap['SnapshotId'])
+        for tag in snap['Tags']:
+            if tag['Key'] != 'KeepForever':
+                skipping_this_one = False
+                continue
+            else:
+                skipping_this_one = True
+
+        if skipping_this_one == True:
+            print "Skipping snapshot %s (marked KeepForever)" % snap['SnapshotId']
+            # do nothing else
+        else:
+            print "Deleting snapshot %s" % snap['SnapshotId']
+            ec.delete_snapshot(SnapshotId=snap['SnapshotId'])
